+++ src/components/RequestInformation/RequestInformation.js
@@ -43,7 +43,9 @@ const RequestInformation = ({
   isTitleLevelRequest,
   isSelectedInstance,
   isSelectedItem,
-  requestTypeError,
+  isSelectedUser,
+  isRequestTypesReceived,
+  isRequestTypeLoading,
   values,
 }) => {
   const isEditForm = isFormEditing(request);
@@ -54,16 +56,20 @@ const RequestInformation = ({
   const isExpirationDate = isEditForm && request.status === requestStatuses.AWAITING_PICKUP;
   const isHoldShelfExpireDate = isEditForm && request.status !== requestStatuses.AWAITING_PICKUP;
   const isItemOrTitleSelected = isTitleLevelRequest ? isSelectedInstance : isSelectedItem;
-  const isRequestTypeDisabled = requestTypeOptions.length === 0 || !isItemOrTitleSelected;
+  const isRequestTypeDisabled = requestTypeOptions.length === 0 || !(isItemOrTitleSelected && isSelectedUser);
   const validateRequestType = useCallback((requestType) => {
-    if (!requestType && isItemOrTitleSelected && requestTypeOptions.length === 0) {
-      return <FormattedMessage id={getNoRequestTypeErrorMessageId(isTitleLevelRequest)} />;
-    } else if (!requestType) {
-      return <FormattedMessage id="ui-requests.errors.requestType.selectItem" />;
+    if (isItemOrTitleSelected && isSelectedUser) {
+      if (requestTypeOptions.length === 0 && isRequestTypesReceived) {
+        return <FormattedMessage id={getNoRequestTypeErrorMessageId(isTitleLevelRequest)} />;
+      }
+
+      if (!requestType && requestTypeOptions.length !== 0) {
+        return <FormattedMessage id="ui-requests.errors.requestType.selectItem" />;
+      }
     }
 
     return undefined;
-  }, [isItemOrTitleSelected, requestTypeOptions, isTitleLevelRequest]);
+  }, [isItemOrTitleSelected, isSelectedUser, requestTypeOptions, isTitleLevelRequest, isRequestTypesReceived]);
 
   return (
     <>
@@ -79,7 +85,7 @@ const RequestInformation = ({
               data-test-request-type
               xs={3}
             >
-              {!requestTypeError && (isEditForm ?
+              {isEditForm ?
                 <KeyValue
                   label={<FormattedMessage id="ui-requests.requestType" />}
                   value={<FormattedMessage id={requestTypesTranslations[request.requestType]} />}
@@ -95,7 +101,9 @@ const RequestInformation = ({
                     input,
                     meta,
                   }) => {
-                    const error = (meta.touched && meta.error) || null;
+                    const selectItemError = requestTypeOptions.length !== 0 && meta.touched && meta.error;
+                    const noAvailableRequestTypesError = !isRequestTypeLoading && isRequestTypesReceived && requestTypeOptions.length === 0 && meta.error;
+                    const error = selectItemError || noAvailableRequestTypesError || null;
 
                     return (
                       <Select
@@ -127,13 +135,7 @@ const RequestInformation = ({
                       </Select>
                     );
                   }}
-                </Field>)
-              }
-              {requestTypeError &&
-                <KeyValue
-                  label={<FormattedMessage id="ui-requests.requestType" />}
-                  value={<FormattedMessage id="ui-requests.noRequestTypesAvailable" />}
-                />
+                </Field>
               }
             </Col>
             <Col xs={2}>
@@ -235,7 +237,6 @@ const RequestInformation = ({
 RequestInformation.propTypes = {
   isTlrEnabledOnEditPage: PropTypes.bool.isRequired,
   MetadataDisplay: PropTypes.func.isRequired,
-  requestTypeError: PropTypes.bool,
   isTitleLevelRequest: PropTypes.bool.isRequired,
   isSelectedInstance: PropTypes.bool.isRequired,
   isSelectedItem: PropTypes.bool.isRequired,