+++ src/utils.js
@@ -31,7 +31,6 @@ import {
   requestStatuses,
   requestTypesMap,
   REQUEST_LEVEL_TYPES,
-  REQUEST_TYPES,
   createModes,
   INVALID_REQUEST_HARDCODED_ID,
   REQUEST_FORM_FIELD_NAMES,
@@ -140,21 +139,6 @@ export function getRequestTypeOptions(item) {
 
 export function isPagedItem(item) {
   return item?.status?.name === itemStatuses.PAGED;
-}
-
-export function hasNonRequestableStatus(item) {
-  return includes([
-    itemStatuses.AGED_TO_LOST,
-    itemStatuses.CLAIMED_RETURNED,
-    itemStatuses.DECLARED_LOST,
-    itemStatuses.INTELLECTUAL_ITEM,
-    itemStatuses.IN_PROCESS_NON_REQUESTABLE,
-    itemStatuses.LONG_MISSING,
-    itemStatuses.LOST_AND_PAID,
-    itemStatuses.UNAVAILABLE,
-    itemStatuses.UNKNOWN,
-    itemStatuses.WITHDRAWN,
-  ], item?.status?.name);
 }
 
 export function isDelivery(request) {
@@ -330,15 +314,6 @@ export const getRequestLevelValue = (value) => {
   return value
     ? REQUEST_LEVEL_TYPES.TITLE
     : REQUEST_LEVEL_TYPES.ITEM;
-};
-
-// can be removed
-export const getInstanceRequestTypeOptions = () => {
-  return [
-    REQUEST_TYPES[requestTypesMap.PAGE],
-    REQUEST_TYPES[requestTypesMap.HOLD],
-    REQUEST_TYPES[requestTypesMap.RECALL],
-  ];
 };
 
 export const getInstanceQueryString = (hrid, id) => `("hrid"=="${hrid}" or "id"=="${id || hrid}")`;